스크립트 수정 2

+++ scripts/before_install.sh
@@ -3,13 +3,19 @@
 sudo yum update -y
 sudo yum install -y docker
 sudo systemctl start docker
-sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
+sudo systemctl enable docker
 
-# 실행 중인 도커 컨테이너 이미지 정지, 삭제
-if [ -x "$(command -v docker-compose)" ]; then
-  docker-compose down
-  docker rmi $(docker images -a -q)
-else
-  echo "docker-compose is not installed"
+# Docker Compose 설치
+if [ ! -f /usr/local/bin/docker-compose ]; then
+    sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
 fi
+
+## 실행 중인 모든 도커 컨테이너 정지 및 제거
+#docker-compose down
+#docker rm $(docker ps -a -q)
+#
+## 사용하지 않는 도커 이미지 삭제
+#docker rmi $(docker images -aq) --force
+
+
